Move entity to its own file
