+++ custom_components/hilo/sensor.py
@@ -1,4 +1,5 @@
 """Support for various Hilo sensors."""
+
 from __future__ import annotations
 
 import asyncio
@@ -6,6 +7,8 @@ from datetime import datetime, timedelta, timezone
 from os.path import isfile
 
 import aiofiles
+import homeassistant.util.dt as dt_util
+import ruyaml as yaml
 from homeassistant.components.integration.sensor import METHOD_LEFT, IntegrationSensor
 from homeassistant.components.sensor import (
     SensorDeviceClass,
@@ -25,6 +28,8 @@ from homeassistant.const import (
     UnitOfPower,
     UnitOfSoundPressure,
     UnitOfTemperature,
+)
+from homeassistant.const import (
     __short_version__ as current_version,
 )
 from homeassistant.core import HomeAssistant
@@ -33,15 +38,13 @@ from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.event import async_track_state_change_event
 from homeassistant.helpers.restore_state import RestoreEntity
 from homeassistant.util import Throttle, slugify
-import homeassistant.util.dt as dt_util
 from packaging.version import Version
 from pyhilo.const import UNMONITORED_DEVICES
 from pyhilo.device import HiloDevice
 from pyhilo.event import Event
 from pyhilo.util import from_utc_timestamp
-import ruyaml as yaml
 
-from . import Hilo, HiloEntity
+from . import Hilo
 from .const import (
     CONF_ENERGY_METER_PERIOD,
     CONF_GENERATE_ENERGY_METERS,
@@ -64,6 +67,7 @@ from .const import (
     TARIFF_LIST,
     WEATHER_CONDITIONS,
 )
+from .entity import HiloEntity
 from .managers import EnergyManager, UtilityManager
 
 WIFI_STRENGTH = {
